Fixed deleting user file with dddd-dd-dd like suffix

+++ lab4/upback.sh
@@ -6,8 +6,9 @@ mkdir -p ~/restore
 rm ~/restore/*
 cp $prevBackup/* ~/restore
 for file in `ls ~/restore`; do
-  if [ `echo $file | grep -E "[0-9]{4}-[0-9]{2}-[0-9]{2}$"` ]; then
+  prefix=`echo $file | sed -E 's/\.[0-9]{4}-[0-9]{2}-[0-9]{2}$//g'`
+  count=`ls ~/restore | grep -c "$prefix"`
+  if [ `echo $file | grep -E "\.[0-9]{4}-[0-9]{2}-[0-9]{2}$"` ] && [ $count -gt 1 ]; then
     rm ~/restore/$file
   fi
 done
-#ls ~/restore | grep -e "[0-9]{4}-[0-9]{2}-[0-9]{2}" | rm
